backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
django-celery-results==2.5.1
channels==4.1.0
channels-redis==4.2.0
pillow==10.3.0
boto3==1.34.101
twilio==9.1.0
gunicorn==22.0.0
whitenoise==6.6.0
